feat: Add prepare_4d_prefix_attention_mask function
The `prepare_4d_prefix_attention_mask` function is added to the `utils.py` file in the `axolotl/monkeypatch` directory. This function prepares a 4D attention mask by applying prefix locations to set columns to 1 where the prefix is present. This function improves the handling of attention masks in certain scenarios.